requirements.txt
flask==1.1.2
flask_sqlalchemy
psycopg2
pandas
